Remove ENVOY_SHA and API_SHA from Dockerfiles
These fields do not actually work -- there have been no Istio releases
that included these. It is also misleading, see
https://github.com/istio/proxy/pull/2364.

Note -- the important fields, that are actually included in the metadata
and sent to pilot, Istio Version and Proxy Version still remain.

This only removes some labels that have not ever been included - they
were introduced in 1.2 and in all 1.2 images they are `""`.