Split up EnvoyFilter logic for cluster
Currently, we build up all of the clusters, then we apply all the
EnvoyFilter patching. This makes caching impossible, as we cannot
independantly cache individual clusters.

This change decomposes the logic: we check if a cluster is removed, and
if not we do any merges. At the end, we insert all newly added clusters.

This enables us to add caching in the future