Optimize sidecar injection
Normally I would think this is the last place to optimize. However, the
jsonpatch library we were using is so slow this ends up being a
bottleneck.

In normal clusters, I regularly see this using up ~10% of memory (for
short period of time, so mostly GC pressure). However, in some cases of
large pod specs, I have observed this taking >10s (webhook timeout) and
max CPU to inject a single pod.

```
before: 956123097 ns/op        345202583 B/op    532391 allocs/op
after:  18518398 ns/op         6066800 B/op      68773 allocs/op
```

Additionally, this reduces the amount of dependencies we pull in which is always nice. The
json-patch library we are using now is what is used by Kubernetes (and other parts of
our project).