Replace iproute2 dependency with native go code
Currently, distroless images do not have iproute2 binary. This means
ipv6 or tproxy are 100% broken on distroless.

One option is to go add the libraries to distroless. However, they are
*massive*, and sort of defeat the purpose of distroless. The
`github.com/vishvananda/netlink` library is already a common library
used by our dependencies (so we are moving from indirect -> direct
dependency) that configures the same underlying NETLINK configurations
that the `ip` command does, but with pure go code in a small dependency.

This updates all of our usages of `ip` to use this library. With this,
tproxy and ipv6 both work with distroless images. While we currently do
not have CI testing of this, I am working on it (this is a prereq).