waypoint: refactor code to reduce duplication
No need to call the creation code in two places. Also, the old usage was
technically unsafe but happens to work in practice -- we are appending
to the same list twice, which could impact itself. I think its always
going to allocate a copy anyways, but its error prone to do so